Move some entity types
