--- crates/modor_graphics/src/internal/mod.rs
@@ -1,4 +1,0 @@
-#[macro_use]
-pub(crate) mod utils;
-
-pub(crate) mod window;
