+++ crates/modor_graphics/src/lib.rs
@@ -4,17 +4,16 @@
 extern crate modor;
 
 #[macro_use]
-mod internal;
+mod utils;
+
 mod appearance;
 mod backend;
-mod background;
 mod module;
 mod runner;
 mod storages;
 mod window;
 
 pub use appearance::*;
-pub use background::*;
 pub use module::*;
 pub use runner::*;
 pub use window::*;
