+++ crates/modor_graphics/src/utils.rs
